Fixed inheritance problem
Now there's a proper inheritance in the for the config classes
