+++ bukkit/src/main/kotlin/com/github/secretx33/secretcfg/bukkit/enumconfig/EnumConfigImpl.kt
@@ -23,13 +23,9 @@
  */
 package com.github.secretx33.secretcfg.bukkit.enumconfig
 
-import com.cryptomorin.xseries.XEnchantment
-import com.cryptomorin.xseries.XMaterial
-import com.cryptomorin.xseries.XPotion
-import com.cryptomorin.xseries.XSound
-import com.github.secretx33.secretcfg.bukkit.extensions.isAir
-import com.github.secretx33.secretcfg.bukkit.serializer.ColorParser
-import com.github.secretx33.secretcfg.bukkit.serializer.ItemSerializer
+import com.github.secretx33.secretcfg.bukkit.config.BukkitConfig
+import com.github.secretx33.secretcfg.bukkit.config.Config
+import com.github.secretx33.secretcfg.bukkit.config.ConfigImpl
 import com.github.secretx33.secretcfg.core.enumconfig.AbstractEnumConfig
 import com.github.secretx33.secretcfg.core.enumconfig.ConfigEnum
 import org.bukkit.*
@@ -51,201 +47,9 @@ class EnumConfigImpl<U> (
     logger: Logger = plugin.logger,
     copyDefault: Boolean = true,
     filePresentInJar: Boolean = true,
-) : AbstractEnumConfig<U>(plugin, plugin.dataFolder, configClass, path, logger, copyDefault, filePresentInJar),
-    EnumConfig<U> where U : ConfigEnum, U : Enum<U> {
-
-    private val colorParser = ColorParser(manager.fileName, logger)
-
-    // ItemStack Serialization & Deserialization
-
-    override fun serialize(key: String, item: ItemStack) = set(key, ItemSerializer.toString(item))
-
-    override fun deserializeItem(key: String): ItemStack? {
-        val item = getString(key, "").takeIf { it.isNotBlank() } ?: return null
-        return ItemSerializer.fromString(item)
-    }
-
-    override fun deserializeItem(key: String, default: Supplier<ItemStack>): ItemStack {
-        val item = getString(key, "").takeIf { it.isNotBlank() } ?: return default.get()
-        return ItemSerializer.fromStringOrNull(item) ?: default.get()
-    }
-
-    override fun serialize(key: String, itemList: List<ItemStack>) {
-        itemList.filter { !it.isAir && it.type.isItem }
-            .map { ItemSerializer.toString(it) }
-            .let { set(key, it) }
-    }
-
-    override fun deserializeItemList(key: String, default: Supplier<List<ItemStack>>): List<ItemStack> {
-        return cachedList(key, default) { list ->
-            list.mapNotNull { ItemSerializer.fromStringOrNull(it) }
-        }
-    }
-
-    // Material
-
-    override fun getMaterial(key: String, default: Material): Material
-            = cachedEnum(key, default) { enum -> XMaterial.matchXMaterial(enum).map { it.parseMaterial() }?.orElse(null) }
-
-    override fun getMaterialList(key: String, default: List<Material>, filter: Predicate<Material>): List<Material> {
-        return filteredCachedList(key, default, filter) { list ->
-            list.mapNotNull { entry ->
-                XMaterial.matchXMaterial(entry).map { it.parseMaterial() }?.orElse(null) ?: run {
-                    warnInvalidEntry(key, entry)
-                    null
-                }
-            }
-        }
-    }
-
-    override fun getMaterialSet(key: String, default: Set<Material>, filter: Predicate<Material>): Set<Material> {
-        return filteredCachedSet(key, default, filter) { list ->
-            list.mapNotNull { entry ->
-                XMaterial.matchXMaterial(entry).map { it.parseMaterial() }?.orElse(null) ?: run {
-                    warnInvalidEntry(key, entry)
-                    null
-                }
-            }
-        }
-    }
-
-    // EntityType
-
-    override fun getEntityType(key: String, default: EntityType): EntityType
-            = cachedEnum(key, default) { type -> EntityType.values().firstOrNull { it.name.equals(type, ignoreCase = true) } }
-
-    override fun getEntityTypeList(key: String, default: List<EntityType>, filter: Predicate<EntityType>): List<EntityType>
-            = getEnumList(key, default, EntityType::class, filter)
-
-    override fun getEntityTypeSet(key: String, default: Set<EntityType>, filter: Predicate<EntityType>): Set<EntityType>
-            = getEnumSet(key, default, EntityType::class, filter)
-
-
-    // Potion Effect Type
-
-    override fun getPotionEffect(key: String, default: PotionEffectType): PotionEffectType
-            = cachedStringBased(key, default) { type -> XPotion.parsePotionEffectFromString(type)?.type }
-
-    override fun getPotionEffectList(key: String, default: List<PotionEffectType>, filter: Predicate<PotionEffectType>): List<PotionEffectType> {
-        return filteredCachedList(key, default, filter) { list ->
-            list.mapNotNull { type ->
-                XPotion.parsePotionEffectFromString(type)?.type ?: run {
-                    warnInvalidEntry(key, type)
-                    null
-                }
-            }
-        }
-    }
-
-    override fun getPotionEffectSet(key: String, default: Set<PotionEffectType>, filter: Predicate<PotionEffectType>): Set<PotionEffectType> {
-        return filteredCachedSet(key, default, filter) { set ->
-            set.mapNotNull { type ->
-                XPotion.parsePotionEffectFromString(type)?.type ?: run {
-                    warnInvalidEntry(key, type)
-                    null
-                }
-            }
-        }
-    }
-
-
-    // Particle
-
-    override fun getParticle(key: String, default: Particle): Particle
-            = cachedEnum(key, default) { enum -> Particle.values().firstOrNull { it.name.equals(enum, ignoreCase = true) } }
-
-    override fun getParticleList(key: String, default: List<Particle>): List<Particle>
-            = getEnumList(key, default, Particle::class)
-
-    override fun getParticleSet(key: String, default: Set<Particle>): Set<Particle>
-            = getEnumSet(key, default, Particle::class)
-
-
-    // Color
-
-    override fun getColor(key: String, default: Color): Color
-            = cachedStringBased(key, default) { color -> colorParser.parse(key, color) }
-
-    // parse the color entries, warning about invalid colors
-    override fun getColorList(key: String, default: List<Color>): List<Color>
-            = cachedList(key, default) { list -> list.mapNotNull { colorParser.parse(key, it) } }
-
-    // parse the color entries, warning about invalid colors
-    override fun getColorSet(key: String, default: Set<Color>): Set<Color>
-            = cachedSet(key, default) { list -> list.mapNotNullTo(mutableSetOf()) { colorParser.parse(key, it) } }
-
-
-    // Dye Color
-
-    override fun getDyeColor(key: String, default: DyeColor): DyeColor
-            = cachedEnum(key, default) { enum -> DyeColor.values().firstOrNull { it.name.equals(enum, ignoreCase = true) }  }
-
-    override fun getDyeColorList(key: String, default: List<DyeColor>): List<DyeColor>
-            = getEnumList(key, default, DyeColor::class)
-
-
-    override fun getDyeColorSet(key: String, default: Set<DyeColor>): Set<DyeColor>
-            = getEnumSet(key, default, DyeColor::class)
-
-    // Enchantment
-
-    override fun getEnchant(key: String, default: Enchantment): Enchantment
-            = cachedStringBased(key, default) { enchant -> XEnchantment.matchXEnchantment(enchant).map { it.parseEnchantment() }?.orElse(null) }
-
-    override fun getEnchantList(key: String, default: List<Enchantment>, filter: Predicate<Enchantment>): List<Enchantment> {
-        return filteredCachedList(key, default, filter) { list ->
-            list.mapNotNull { enchant ->
-                XEnchantment.matchXEnchantment(enchant).map { it.parseEnchantment() }?.orElse(null) ?: run {
-                    warnInvalidEntry(key, enchant)
-                    null
-                }
-            }
-        }
-    }
-
-    override fun getEnchantSet(key: String, default: Set<Enchantment>, filter: Predicate<Enchantment>): Set<Enchantment> {
-        return filteredCachedSet(key, default, filter) { set ->
-            set.mapNotNull { enchant ->
-                XEnchantment.matchXEnchantment(enchant).map { it.parseEnchantment() }?.orElse(null) ?: run {
-                    warnInvalidEntry(key, enchant)
-                    null
-                }
-            }
-        }
-    }
-
-    override fun getItemFlags(key: String, default: Set<ItemFlag>, filter: Predicate<ItemFlag>): Set<ItemFlag>
-        = getEnumSet(key, default, ItemFlag::class, filter)
-
-    @Suppress("UNCHECKED_CAST")
-    override fun getSound(key: String): Triple<Sound, Float, Float>? {
-        cache[key]?.let { runCatching { it as Triple<Sound, Float, Float> }.getOrNull()?.let { return it } }
-
-        val sounds = manager.getString(key, "").split(':', limit = 3).map { it.replace(" ", "") }
-        // no sound
-        if(sounds.isEmpty() || sounds[0].isBlank()) return null
-
-        // for "SOUND"
-        val sound = XSound.matchXSound(sounds[0]).map { it.parseSound() }?.orElse(null) ?: run {
-            warnInvalidEntry(key, sounds[0])
-            return null
-        }
-        if(sounds.size == 1) return Triple(sound, 1f, 1f).also { cache[key] = it }
-
-        // for "SOUND:1.0" as in "SOUND:VOLUME"
-        val volume = sounds[1].toDoubleOrNull()?.coerceAtLeast(0.0) ?: run {
-            logger.warning("Could not parse volume '${sounds[1]}' of sound '${sound.name}' of key '$key' in file '$file', please fix your configurations and reload.")
-            return Triple(sound, 1f, 1f).also { cache[key] = it }
-        }
-        if(sounds.size == 2) return Triple(sound, volume.toFloat(), 1f).also { cache[key] = it }
-
-        // for "SOUND:1.0:1.0" as in "SOUND:VOLUME:PITCH"
-        val pitch = sounds[2].toDoubleOrNull()?.coerceAtLeast(0.01) ?: run {
-            logger.warning("Could not parse pitch '${sounds[2]}' of sound '${sound.name}' of key '$key' in file '$file', please fix your configurations and reload.")
-            return Triple(sound, volume.toFloat(), 1f).also { cache[key] = it }
-        }
-        return Triple(sound, volume.toFloat(), pitch.toFloat()).also { cache[key] = it }
-    }
+    private val bukkitConfig: Config = ConfigImpl(plugin, path, logger, copyDefault, filePresentInJar),
+) : AbstractEnumConfig<U>(configClass, bukkitConfig),
+    BukkitConfig by bukkitConfig, EnumConfig<U> where U : ConfigEnum, U : Enum<U> {
 
     override fun serialize(key: U, item: ItemStack) = serialize(key.path, item)
